Separating the emails from webform submissions.
This is so we can send different emails to the two groups.
The email that is specified in the webform edit page gets
the typical webform submission email, and the webform
submission email addresses get the email with the file
attachment.

Still need to format the email to the submission emails.

TODO: Figure out how to record submissions emails into the
submission, while disallowing the sending of default emails.
Right now email fields in the webform do not show up in the
submissions results.