Fixed issues with the webform submission data.
Now using hook_mail_alter() to modify the email that is being sent
to the user.  This allows us to leave all form values in tact so
they can be saved to the webform result listing.  We should be able
to accomplish the same thing as before with the separation of the
emails between the email addresses set in the webform edit page
and the submission form values that contain email addresses.

Now that I have that squared away, on to formatting the actual
email.